notebookOrKernelLanguageTelemetry
